refactoring utils into tagproInterface and asserts
